Added python scripts for comparing two datasets
The comparison files are placed in the python_scripts\python_scripts_datacompare folder.
TransposonRead_Profile_Compare.py creates a barplot for two datasets for visual comparison of the insertion locations and amount of reads for each transposon.
statistics_perchromosome_Compare.py creates a text file that shows a numerical overview of the two datasets by ways of calculating some stastical values of the datasets.